refactor: Remove unused code

+++ app/routers/routes.py
@@ -52,10 +52,11 @@ def routes(
         
     @router.delete("/{id}", response_model=None)
     async def delete(id: str):
-        # item_to_delete = await get_func(id)
-        # if not item_to_delete:
-        #     raise CustomValidationException(status_code=204, detail="Item does not exist.")
-        return await delete_func(id)
-        #return {"detail": "Item deleted successfully"}
+        item_to_delete = await get_func(id)
+        if not item_to_delete:
+            raise HTTPException(status_code=404, detail="Item not found")
+        await delete_func(id)
+        return {"detail": "Item deleted successfully"}
+
 
     return router
